# openapi_patched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/carts": {
      "get": {
        "summary": "获取购物车商品列表",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口（后端服务）\n【前置条件】需要用户登录\n【鉴权】Bearer Token",
        "operationId": "获取购物车商品列表",
        "tags": [
          "购物车管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT 令牌，格式：Bearer {token}",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartsListResponse"
                },
                "example": {
                  "code": 200,
                  "data": [
                    {
                      "id": 1,
                      "productId": 1001,
                      "productName": "iPhone 15 Pro",
                      "skuId": 2001,
                      "productImage": "https://example.com/images/iphone15pro.jpg",
                      "skuName": "深空黑色 256GB",
                      "price": 8999,
                      "quantity": 1,
                      "subtotal": 8999
                    },
                    {
                      "id": 2,
                      "productId": 1002,
                      "productImage": "https://example.com/images/airpods-pro2.jpg",
                      "productName": "AirPods Pro 2",
                      "skuId": 2002,
                      "price": 1899,
                      "skuName": "USB-C 充电盒",
                      "quantity": 2,
                      "subtotal": 3798
                    }
                  ],
                  "message": "操作成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "购物车模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441868511-run"
      },
      "post": {
        "summary": "添加商品到购物车",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】将商品加入当前用户的购物车",
        "operationId": "添加商品到购物车",
        "tags": [
          "购物车管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartsRequest"
              },
              "example": {
                "productId": 2005,
                "quantity": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCartsResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "cartTotal": 3,
                    "id": 1002
                  },
                  "message": "添加成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "购物车模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441727940-run"
      },
      "delete": {
        "summary": "清空购物车",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要用户登录\n【功能描述】清空当前用户的整个购物车内容",
        "operationId": "清空购物车",
        "tags": [
          "购物车管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCartResponse"
                },
                "example": {
                  "code": 200,
                  "message": "购物车已清空"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "购物车模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441879013-run"
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "标准错误响应",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "details": {
            "type": "object",
            "description": "详细信息",
            "x-apifox-orders": [],
            "properties": {},
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "code",
          "message"
        ],
        "x-apifox-orders": [
          "code",
          "message",
          "details"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "CartsListResponse": {
        "properties": {
          "code": {
            "description": "响应状态码",
            "type": "integer"
          },
          "message": {
            "description": "响应消息",
            "type": "string"
          },
          "data": {
            "description": "购物车商品列表",
            "items": {
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "购物车项 ID"
                },
                "productId": {
                  "description": "商品 ID",
                  "type": "integer"
                },
                "skuId": {
                  "type": "integer",
                  "description": "SKU ID"
                },
                "productImage": {
                  "type": "string",
                  "description": "商品图片 URL"
                },
                "productName": {
                  "description": "商品名称",
                  "type": "string"
                },
                "price": {
                  "type": "number",
                  "description": "商品单价"
                },
                "skuName": {
                  "description": "SKU 规格名称",
                  "type": "string"
                },
                "quantity": {
                  "description": "购买数量",
                  "type": "integer"
                },
                "subtotal": {
                  "description": "小计金额",
                  "type": "number"
                }
              },
              "type": "object",
              "required": [
                "id",
                "productId",
                "skuId",
                "productName",
                "price",
                "quantity",
                "subtotal"
              ],
              "x-apifox-orders": [
                "id",
                "productId",
                "skuId",
                "productImage",
                "productName",
                "price",
                "skuName",
                "quantity",
                "subtotal"
              ],
              "x-apifox-ignore-properties": []
            },
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "x-apifox-orders": [
          "code",
          "message",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "DeleteCartResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "CreateCartsRequest": {
        "properties": {
          "productId": {
            "description": "要添加的商品ID",
            "type": "integer"
          },
          "quantity": {
            "description": "添加数量",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "quantity"
        ],
        "type": "object",
        "x-apifox-orders": [
          "productId",
          "quantity"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "CreateCartsResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "添加后的相关状态",
            "properties": {
              "cartTotal": {
                "description": "购物车当前商品总数",
                "type": "integer"
              },
              "id": {
                "description": "购物车条目ID",
                "type": "integer"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "cartTotal",
              "id"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}